feat(NGUI-111) UI agent performance stats from the evaluation
