+++ tests/ai_eval_components/eval.py
@@ -1,15 +1,30 @@
 import asyncio
-import getopt
 import json
 import os
-import sys
-from io import TextIOWrapper
-from pathlib import Path
-from typing import Any
+import time
 
+from ai_eval_components.eval_perfstats import print_perf_stats, report_perf_stats
+from ai_eval_components.eval_reporting import (
+    ERR_FILE_SUFFIX,
+    console_print_progress_dot,
+    get_errors_dir,
+    get_llm_output_dir,
+    init_agent_out_stream,
+    print_stats,
+    report_err_dataset,
+    report_err_sys,
+    report_err_uiagent,
+    report_success,
+)
+from ai_eval_components.eval_utils import (
+    assert_array_not_empty,
+    assert_str_not_blank,
+    get_dataset_files,
+    load_args,
+    load_dataset_file,
+    validate_dataset_row,
+)
 from ai_eval_components.types import (
-    BASE_DATASET_PATH,
-    BASE_MODULE_PATH,
     DATASET_FILE_SUFFIX,
     DatasetRow,
     DatasetRowAgentEvalResult,
@@ -27,8 +42,6 @@ from next_gen_ui_llama_stack.llama_stack_inference import LlamaStackAgentInferen
 INFERENCE_MODEL_DEFAULT = "granite3.1-dense:2b"
 LLAMA_STACK_HOST_DEFAULT = "localhost"
 LLAMA_STACK_PORT_DEFAULT = "5001"
-
-ERR_FILE_SUFFIX = "-errors.txt"
 
 
 def init_inference() -> InferenceBase:
@@ -88,63 +101,17 @@ def check_result_explicit(
             # TODO component specific evals where fields have to be of some type, eg link to image for image component, link to video etc
 
 
-def assert_array_not_empty(
-    value_object,
-    value_name: str,
-    error_code: str,
-    errors: list[EvalError],
-    err_msg=None,
-):
-    """Assert that array value expected under `value_name` in the `value_object` is not empty, add `error_code` into `errors` if it is empty and return `False`"""
-    try:
-        value = value_object[value_name]
-
-        if not value or len(value) == 0:
-            errors.append(
-                EvalError(error_code, err_msg if err_msg else f"array is '{value}'")
-            )
-            return False
-        else:
-            return True
-    except KeyError:
-        errors.append(EvalError(error_code, err_msg if err_msg else "array is missing"))
-        return False
-
-
-def assert_str_not_blank(
-    value_object,
-    value_name: str,
-    error_code: str,
-    errors: list[EvalError],
-    err_msg=None,
-):
-    """Assert that string value expected under `value_name` in the `value_object` is not a blank string, add `error_code` into `errors` if it is blank and return `False`"""
-    try:
-        value = value_object[value_name]
-        if not value or len(value.strip()) == 0:
-            errors.append(
-                EvalError(
-                    error_code, err_msg if err_msg else f"string value is '{value}'"
-                )
-            )
-            return False
-        else:
-            return True
-    except KeyError:
-        errors.append(
-            EvalError(error_code, err_msg if err_msg else "string is missing")
-        )
-        return False
-
-
 def evaluate_agent_for_dataset_row(dsr: DatasetRow, inference: InferenceBase):
+    """Run agent evaluation for one dataset row"""
     errors: list[EvalError] = []
     input_data = InputData(id="myid", data=dsr["backend_data"])
 
     # separate steps so we can see LLM response even if it is invalid JSON
+    time_start = round(time.time() * 1000)
     llm_response = asyncio.run(
         component_selection_inference(dsr["user_prompt"], inference, input_data)
     )
+    report_perf_stats(time_start, round(time.time() * 1000), dsr["expected_component"])
 
     try:
         component: UIComponentMetadata = json.loads(llm_response)
@@ -166,211 +133,6 @@ def evaluate_agent_for_dataset_row(dsr: DatasetRow, inference: InferenceBase):
     return DatasetRowAgentEvalResult(llm_response, errors)
 
 
-def validate_dataset_row(dsr: DatasetRow):
-    errors: list[EvalError] = []
-    assert_str_not_blank(dsr, "id", "id.missing", errors)
-    assert_str_not_blank(
-        dsr, "expected_component", "expected_component.missing", errors
-    )
-    assert_str_not_blank(dsr, "user_prompt", "user_prompt.missing", errors)
-    if assert_str_not_blank(dsr, "backend_data", "backend_data.missing", errors):
-        try:
-            json.loads(dsr["backend_data"])
-        except json.JSONDecodeError as e:
-            errors.append(EvalError("backend_data.invalid_json", str(e)))
-    return errors
-
-
-def load_dataset_file(filepath: Path):
-    dataset: list[DatasetRow]
-
-    with filepath.open("r") as f:
-        dataset = json.load(f)
-
-    return dataset
-
-
-def console_handle_progress_dot(is_dot: bool):
-    if is_dot:
-        print()
-    return False
-
-
-def get_dataset_dir():
-    dataset_dir_path_config = os.getenv("DATASET_DIR")
-    if dataset_dir_path_config:
-        dataset_dir_path = Path(dataset_dir_path_config)
-    else:
-        dataset_dir_path = Path.cwd() / BASE_DATASET_PATH
-    print(f"Loading dataset from folder: {dataset_dir_path}")
-    return dataset_dir_path
-
-
-def get_dataset_files(filename: str):
-    dataset_dir_path = get_dataset_dir()
-    if filename:
-        dataset_files = [dataset_dir_path / filename]
-    else:
-        dataset_files = [
-            f
-            for f in dataset_dir_path.iterdir()
-            if f.is_file() and f.match("*" + DATASET_FILE_SUFFIX)
-        ]
-        dataset_files.sort()
-    return dataset_files
-
-
-def get_errors_dir():
-    errors_dir_path_config = os.getenv("ERRORS_DIR")
-    if errors_dir_path_config:
-        errors_dir_path = Path(errors_dir_path_config)
-    else:
-        errors_dir_path = Path.cwd() / (BASE_MODULE_PATH + "errors/")
-    if not errors_dir_path.exists():
-        errors_dir_path.mkdir(parents=True)
-    else:
-        [
-            f.unlink()
-            for f in errors_dir_path.iterdir()
-            if f.is_file() and f.match("*" + ERR_FILE_SUFFIX)
-        ]
-    print(f"Writing errors into folder: {errors_dir_path}")
-    return errors_dir_path
-
-
-def get_llm_output_dir():
-    llm_output_dir_path = Path.cwd() / (BASE_MODULE_PATH + "llm_out/")
-    if not llm_output_dir_path.exists():
-        llm_output_dir_path.mkdir(parents=True)
-    else:
-        [f.unlink() for f in llm_output_dir_path.iterdir() if f.is_file()]
-    print(f"Writing successfull LLM outputs into folder: {llm_output_dir_path}")
-    return llm_output_dir_path
-
-
-def load_args():
-    arg_ui_component = None
-    arg_dataset_file = None
-    arg_write_llm_output = False
-    opts, args = getopt.getopt(sys.argv[1:], "hwc:f:")
-    for opt, arg in opts:
-        if opt == "-h":
-            print("eval.py <arguments>")
-            print("\nArguments:")
-            print(" -c <ui-component-name> - evaluate only named UI component")
-            print(
-                " -f <dataset-file-name> - run only evaluations from the named dataset file"
-            )
-            print(
-                " -w - if present then LLM outputs with successful checks are written into files in 'llm_out' directory"
-            )
-            print(" -h - help")
-            sys.exit()
-        elif opt in ("-c"):
-            arg_ui_component = arg
-        elif opt in ("-w"):
-            arg_write_llm_output = True
-        elif opt in ("-f"):
-            arg_dataset_file = arg
-
-    if arg_ui_component:
-        print(f"Running evaluations for UI component '{arg_ui_component}' ...")
-    else:
-        print("Running evaluations for all UI components ...")
-
-    return arg_ui_component, arg_write_llm_output, arg_dataset_file
-
-
-def report_err_sys(f_err: TextIOWrapper, id: str, e: Exception):
-    print(f"{id} - ERROR SYSTEM {type(e).__name__} - {e}")
-    f_err.write(f"==== SYSTEM {id} ====\n")
-    f_err.write(f"{type(e).__name__} - {e}\n")
-    f_err.write("\n\n")
-    f_err.flush()
-
-
-def report_llm_output(
-    f_out: TextIOWrapper,
-    id: str,
-    eval_result: DatasetRowAgentEvalResult,
-    dsr: DatasetRow,
-):
-    f_out.write(f"==== DATASET ID {id} ====\n")
-    f_out.write("Prompt:\n")
-    f_out.write(dsr["user_prompt"])
-    f_out.write("\nAgent output:\n")
-    json.dump(
-        json.loads(eval_result.llm_output), f_out, indent=2, separators=(",", ": ")
-    )
-    if "src" in dsr and "data_file" in dsr["src"]:
-        ov = dsr["src"]["data_file"]
-        f_out.write(f"\nData file in dataset_src: {ov}")
-    f_out.write("\n\n")
-    f_out.flush()
-
-
-def report_err_uiagent(
-    f_err: TextIOWrapper,
-    id: str,
-    eval_result: DatasetRowAgentEvalResult,
-    dsr: DatasetRow,
-):
-    print(f"{id} - ERROR UIAGENT {eval_result.errors}")
-    f_err.write(f"==== AGENT {id} ====\n")
-    json.dump(
-        json.loads(str(eval_result.errors)), f_err, indent=2, separators=(",", ": ")
-    )
-    f_err.write("\nPrompt:\n")
-    f_err.write(dsr["user_prompt"])
-    f_err.write("\nAgent output:\n")
-    try:
-        json.dump(
-            json.loads(eval_result.llm_output), f_err, indent=2, separators=(",", ": ")
-        )
-    except Exception:
-        # write LLM output even if it is an invalid JSON
-        f_err.write(eval_result.llm_output)
-    if "src" in dsr and "data_file" in dsr["src"]:
-        ov = dsr["src"]["data_file"]
-        f_err.write(f"\nData file in dataset_src: {ov}")
-    f_err.write("\n\n")
-    f_err.flush()
-
-
-def report_err_dataset(f_err: TextIOWrapper, id: str, ds_errors: list[EvalError]):
-    print(f"{id} - ERROR DATASET {ds_errors}")
-    f_err.write(f"==== DATASET {id} ====\n")
-    json.dump(json.loads(str(ds_errors)), f_err, indent=2, separators=(",", ": "))
-    f_err.write("\n\n")
-    f_err.flush()
-
-
-def get_stats_by_component(eval_stats_by_component: dict, component: str):
-    if component not in eval_stats_by_component:
-        eval_stats_by_component[component] = {
-            "num_evals": 0,
-            "num_err_agent": 0,
-            "num_err_system": 0,
-        }
-    return eval_stats_by_component[component]
-
-
-def stats_by_component_add_success(eval_stats_by_component: dict, component: str):
-    sbc = get_stats_by_component(eval_stats_by_component, component)
-    sbc["num_evals"] += 1
-    return sbc
-
-
-def stats_by_component_add_err_agent(eval_stats_by_component: dict, component: str):
-    sbc = stats_by_component_add_success(eval_stats_by_component, component)
-    sbc["num_err_agent"] += 1
-
-
-def stats_by_component_add_err_system(eval_stats_by_component: dict, component: str):
-    sbc = stats_by_component_add_success(eval_stats_by_component, component)
-    sbc["num_err_system"] += 1
-
-
 if __name__ == "__main__":
     arg_ui_component, arg_write_llm_output, arg_dataset_file = load_args()
     errors_dir_path = get_errors_dir()
@@ -378,13 +140,6 @@ if __name__ == "__main__":
         llm_output_dir_path = get_llm_output_dir()
     dataset_files = get_dataset_files(arg_dataset_file)
     inference = init_inference()
-
-    num_evals = 0
-    num_err_ds = 0
-    num_err_agent = 0
-    num_err_system = 0
-
-    eval_stats_by_component: dict[str, Any] = {}
 
     for dataset_file in dataset_files:
         dataset = load_dataset_file(dataset_file)
@@ -402,62 +157,38 @@ if __name__ == "__main__":
             errors_dir_path
             / dataset_file.name.replace(DATASET_FILE_SUFFIX, ERR_FILE_SUFFIX)
         ).open("a") as f_err:
-            if arg_write_llm_output:
-                f_out = (
-                    llm_output_dir_path
-                    / dataset_file.name.replace(DATASET_FILE_SUFFIX, ".txt")
-                ).open("a")
+            f_out = init_agent_out_stream(
+                arg_write_llm_output, llm_output_dir_path, dataset_file
+            )
             i = 0
             is_progress_dot = False
             for dsr in dataset:
-                num_evals += 1
-                id = dsr["id"]
-
-                # show progress
-                i += 1
-                if i % 2 == 0:
-                    print(".", end="", flush=True)
-                    is_progress_dot = True
-
+                i, is_progress_dot = console_print_progress_dot(i, is_progress_dot)
                 try:
                     ds_errors = validate_dataset_row(dsr)
 
                     if len(ds_errors) > 0:
-                        is_progress_dot = console_handle_progress_dot(is_progress_dot)
-                        report_err_dataset(f_err, id, ds_errors)
-                        num_err_ds += 1
+                        is_progress_dot = report_err_dataset(
+                            f_err, ds_errors, dsr, is_progress_dot
+                        )
                     else:
                         eval_result = evaluate_agent_for_dataset_row(dsr, inference)
                         if len(eval_result.errors) > 0:
-                            is_progress_dot = console_handle_progress_dot(
-                                is_progress_dot
+                            is_progress_dot = report_err_uiagent(
+                                f_err, eval_result, dsr, is_progress_dot
                             )
-                            report_err_uiagent(f_err, id, eval_result, dsr)
-                            stats_by_component_add_err_agent(
-                                eval_stats_by_component, dsr["expected_component"]
-                            )
-                            num_err_agent += 1
                         else:
-                            stats_by_component_add_success(
-                                eval_stats_by_component, dsr["expected_component"]
+                            report_success(
+                                f_out, eval_result, dsr, arg_write_llm_output
                             )
-                            if arg_write_llm_output:
-                                report_llm_output(f_out, id, eval_result, dsr)
 
                 except Exception as e:
-                    report_err_sys(f_err, id, e)
-                    stats_by_component_add_err_system(
-                        eval_stats_by_component, dsr["expected_component"]
-                    )
-                    num_err_system += 1
+                    is_progress_dot = report_err_sys(f_err, e, dsr, is_progress_dot)
 
             if arg_write_llm_output:
                 f_out.close()
 
     print("\n\nEvaluations finished:")
-    print(f"Dataset items evalueated: {num_evals}")
-    print(f"Agent errors: {num_err_agent}")
-    print(f"System errors: {num_err_system}")
-    print(f"Dataset errors: {num_err_ds}")
-    print("Results by component:")
-    json.dump(eval_stats_by_component, sys.stdout, indent=2, separators=(",", ": "))
+    print_stats()
+    print("")
+    print_perf_stats()
