+++ tests/ai_eval_components/eval_reporting.py
@@ -1,0 +1,216 @@
+import json
+import os
+import sys
+from io import TextIOWrapper
+from pathlib import Path
+from typing import Any
+
+from ai_eval_components.types import (
+    BASE_MODULE_PATH,
+    DATASET_FILE_SUFFIX,
+    DatasetRow,
+    DatasetRowAgentEvalResult,
+    EvalError,
+)
+
+ERR_FILE_SUFFIX = "-errors.txt"
+
+eval_stats = {
+    "num_evals": 0,
+    "num_err_ds": 0,
+    "num_err_agent": 0,
+    "num_err_system": 0,
+}
+eval_stats_by_component: dict[str, Any] = {}
+
+
+def report_success(
+    f_out: TextIOWrapper,
+    eval_result: DatasetRowAgentEvalResult,
+    dsr: DatasetRow,
+    arg_write_llm_output: bool,
+):
+    """Report successfull eval"""
+    eval_stats["num_evals"] += 1
+    stats_by_component_add_success(dsr["expected_component"])
+    if arg_write_llm_output:
+        id = dsr["id"]
+        f_out.write(f"==== DATASET ID {id} ====\n")
+        f_out.write("Prompt:\n")
+        f_out.write(dsr["user_prompt"])
+        f_out.write("\nAgent output:\n")
+        json.dump(
+            json.loads(eval_result.llm_output), f_out, indent=2, separators=(",", ": ")
+        )
+        if "src" in dsr and "data_file" in dsr["src"]:
+            ov = dsr["src"]["data_file"]
+            f_out.write(f"\nData file in dataset_src: {ov}")
+        f_out.write("\n\n")
+        f_out.flush()
+
+
+def report_err_dataset(
+    f_err: TextIOWrapper,
+    ds_errors: list[EvalError],
+    dsr: DatasetRow,
+    is_progress_dot: bool,
+):
+    """Report dataset error"""
+    eval_stats["num_evals"] += 1
+    eval_stats["num_err_ds"] += 1
+
+    is_progress_dot = console_handle_progress_dot(is_progress_dot)
+    id = dsr["id"]
+    print(f"{id} - ERROR DATASET {ds_errors}")
+    f_err.write(f"==== DATASET {id} ====\n")
+    json.dump(json.loads(str(ds_errors)), f_err, indent=2, separators=(",", ": "))
+    f_err.write("\n\n")
+    f_err.flush()
+
+    return is_progress_dot
+
+
+def report_err_uiagent(
+    f_err: TextIOWrapper,
+    eval_result: DatasetRowAgentEvalResult,
+    dsr: DatasetRow,
+    is_progress_dot: bool,
+):
+    """Report UI Agent error"""
+    eval_stats["num_evals"] += 1
+    stats_by_component_add_err_agent(dsr["expected_component"])
+    eval_stats["num_err_agent"] += 1
+
+    is_progress_dot = console_handle_progress_dot(is_progress_dot)
+    id = dsr["id"]
+    print(f"{id} - ERROR UIAGENT {eval_result.errors}")
+    f_err.write(f"==== AGENT {id} ====\n")
+    json.dump(
+        json.loads(str(eval_result.errors)), f_err, indent=2, separators=(",", ": ")
+    )
+    f_err.write("\nPrompt:\n")
+    f_err.write(dsr["user_prompt"])
+    f_err.write("\nAgent output:\n")
+    try:
+        json.dump(
+            json.loads(eval_result.llm_output), f_err, indent=2, separators=(",", ": ")
+        )
+    except Exception:
+        # write LLM output even if it is an invalid JSON
+        f_err.write(eval_result.llm_output)
+    if "src" in dsr and "data_file" in dsr["src"]:
+        ov = dsr["src"]["data_file"]
+        f_err.write(f"\nData file in dataset_src: {ov}")
+    f_err.write("\n\n")
+    f_err.flush()
+
+    return is_progress_dot
+
+
+def report_err_sys(
+    f_err: TextIOWrapper, e: Exception, dsr: DatasetRow, is_progress_dot: bool
+):
+    """Report system error/exception"""
+    eval_stats["num_evals"] += 1
+    stats_by_component_add_err_system(dsr["expected_component"])
+    eval_stats["num_err_system"] += 1
+
+    is_progress_dot = console_handle_progress_dot(is_progress_dot)
+    id = dsr["id"]
+    print(f"{id} - ERROR SYSTEM {type(e).__name__} - {e}")
+    f_err.write(f"==== SYSTEM {id} ====\n")
+    f_err.write(f"{type(e).__name__} - {e}\n")
+    f_err.write("\n\n")
+    f_err.flush()
+
+    return is_progress_dot
+
+
+def print_stats():
+    """Print eval stats to the stdout"""
+    print(f"Dataset items evalueated: {eval_stats['num_evals']}")
+    print(f"Agent errors: {eval_stats['num_err_agent']}")
+    print(f"System errors: {eval_stats['num_err_system']}")
+    print(f"Dataset errors: {eval_stats['num_err_ds']}")
+    print("Results by component:")
+    json.dump(eval_stats_by_component, sys.stdout, indent=2, separators=(",", ": "))
+    print("")
+
+
+def init_agent_out_stream(
+    arg_write_llm_output: bool, llm_output_dir_path: Path, dataset_file: Path
+):
+    if arg_write_llm_output:
+        return (
+            llm_output_dir_path / dataset_file.name.replace(DATASET_FILE_SUFFIX, ".txt")
+        ).open("a")
+    else:
+        return None
+
+
+def get_stats_by_component(component: str):
+    if component not in eval_stats_by_component:
+        eval_stats_by_component[component] = {
+            "num_evals": 0,
+            "num_err_agent": 0,
+            "num_err_system": 0,
+        }
+    return eval_stats_by_component[component]
+
+
+def stats_by_component_add_success(component: str):
+    sbc = get_stats_by_component(component)
+    sbc["num_evals"] += 1
+    return sbc
+
+
+def stats_by_component_add_err_agent(component: str):
+    sbc = stats_by_component_add_success(component)
+    sbc["num_err_agent"] += 1
+
+
+def stats_by_component_add_err_system(component: str):
+    sbc = stats_by_component_add_success(component)
+    sbc["num_err_system"] += 1
+
+
+def console_print_progress_dot(i: int, is_dot: bool):
+    i += 1
+    if i % 2 == 0:
+        print(".", end="", flush=True)
+        is_dot = True
+    return i, is_dot
+
+
+def console_handle_progress_dot(is_dot: bool):
+    if is_dot:
+        print()
+    return False
+
+
+def get_llm_output_dir():
+    llm_output_dir_path = Path.cwd() / (BASE_MODULE_PATH + "llm_out/")
+    if not llm_output_dir_path.exists():
+        llm_output_dir_path.mkdir(parents=True)
+    else:
+        [f.unlink() for f in llm_output_dir_path.iterdir() if f.is_file()]
+    print(f"Writing successfull LLM outputs into folder: {llm_output_dir_path}")
+    return llm_output_dir_path
+
+
+def get_errors_dir():
+    errors_dir_path_config = os.getenv("ERRORS_DIR")
+    if errors_dir_path_config:
+        errors_dir_path = Path(errors_dir_path_config)
+    else:
+        errors_dir_path = Path.cwd() / (BASE_MODULE_PATH + "errors/")
+    if not errors_dir_path.exists():
+        errors_dir_path.mkdir(parents=True)
+    else:
+        [
+            f.unlink()
+            for f in errors_dir_path.iterdir()
+            if f.is_file() and f.match("*" + ERR_FILE_SUFFIX)
+        ]
+    print(f"Writing errors into folder: {errors_dir_path}")
+    return errors_dir_path
